fix repeat download

+++ crates/snm_shim/src/package_manager_shim.rs
@@ -40,29 +40,35 @@ pub async fn package_manager(prefix: &str, bin_name: &str) -> Result<(), SnmErro
         if package_manager.name == prefix {
             let version = package_manager.version;
 
-            let download_url = snm_package_manage.get_download_url(&version);
+            if snm_package_manage
+                .get_anchor_file_path_buf(&version)?
+                .exists()
+                .not()
+            {
+                let download_url = snm_package_manage.get_download_url(&version);
 
-            let downloaded_file_path_buf =
-                snm_package_manage.get_downloaded_file_path_buf(&version)?;
+                let downloaded_file_path_buf =
+                    snm_package_manage.get_downloaded_file_path_buf(&version)?;
 
-            DownloadBuilder::new()
-                .retries(3)
-                .timeout(
-                    snm_package_manage
-                        .get_snm_config()
-                        .get_download_timeout_secs(),
-                )
-                .write_strategy(WriteStrategy::WriteAfterDelete)
-                .download(&download_url, &downloaded_file_path_buf)
-                .await?;
+                DownloadBuilder::new()
+                    .retries(3)
+                    .timeout(
+                        snm_package_manage
+                            .get_snm_config()
+                            .get_download_timeout_secs(),
+                    )
+                    .write_strategy(WriteStrategy::WriteAfterDelete)
+                    .download(&download_url, &downloaded_file_path_buf)
+                    .await?;
 
-            let runtime_dir_path_buf = snm_package_manage.get_runtime_dir_path_buf(&version)?;
+                let runtime_dir_path_buf = snm_package_manage.get_runtime_dir_path_buf(&version)?;
 
-            snm_package_manage
-                .decompress_download_file(&downloaded_file_path_buf, &runtime_dir_path_buf)?;
+                snm_package_manage
+                    .decompress_download_file(&downloaded_file_path_buf, &runtime_dir_path_buf)?;
 
-            if let Some(parent) = downloaded_file_path_buf.parent() {
-                fs::remove_dir_all(parent)?;
+                if let Some(parent) = downloaded_file_path_buf.parent() {
+                    fs::remove_dir_all(parent)?;
+                }
             }
 
             let binary = snm_package_manage.get_runtime_binary_dir_string(version.as_str())?;
